Move EjectZ from cirq.google to cirq.optimizers
