+++ cirq/circuits/merge_interactions_test.py
@@ -27,7 +27,7 @@ def assert_optimizes(before: cirq.Circuit, expected: cirq.Circuit):
     followup_optimizations = [
         cg.MergeRotations(),
         cg.EjectFullW(),
-        cg.EjectZ(),
+        cirq.EjectZ(),
         cirq.DropNegligible(),
         cirq.DropEmptyMoments()
     ]
