+++ cirq/optimizers/eject_z_test.py
@@ -15,18 +15,16 @@
 from typing import Iterable
 
 import cirq
-import cirq.google as cg
-from cirq.google.eject_z import _try_get_known_z_half_turns
+from cirq.optimizers.eject_z import _try_get_known_z_half_turns
 
 
 def assert_optimizes(before: cirq.Circuit,
                      expected: cirq.Circuit,
-                     pre_opts: Iterable[cirq.OptimizationPass] = (
-                             cg.ConvertToXmonGates(ignore_failures=True),),
+                     pre_opts: Iterable[cirq.OptimizationPass] = (),
                      post_opts: Iterable[cirq.OptimizationPass] = (
-                             cg.ConvertToXmonGates(ignore_failures=True),
-                             cirq.DropEmptyMoments())):
-    opt = cg.EjectZ()
+                             cirq.DropEmptyMoments(),
+                     )):
+    opt = cirq.EjectZ()
 
     if cirq.has_unitary(before):
         cirq.testing.assert_circuits_with_terminal_measurements_are_equivalent(
@@ -48,7 +46,7 @@ def assert_optimizes(before: cirq.Circuit,
 
 
 def assert_removes_all_z_gates(circuit: cirq.Circuit):
-    opt = cg.EjectZ()
+    opt = cirq.EjectZ()
     optimized = circuit.copy()
     opt.optimize_circuit(optimized)
     has_z = any(_try_get_known_z_half_turns(op) is not None
@@ -105,9 +103,7 @@ def test_early_z():
             cirq.Moment([cirq.Z(q)**0.5]),
             cirq.Moment(),
             cirq.Moment(),
-        ]),
-        pre_opts=[cg.ConvertToXmonGates(ignore_failures=True)],
-        post_opts=[cg.ConvertToXmonGates(ignore_failures=True)])
+        ]))
 
 
 def test_multi_z_merges():
